requirements.txt
Django~=5.0
psycopg2-binary~=2.9.9
channels~=3.0.4
python-dotenv~=1.0.1
django-cors-headers~=4.3.1
channels_redis~=4.2.0
djangorestframework~=3.14.0
Pillow~=10.2.0
python-dotenv~=1.0.1
requests~=2.26.0